// frontend/src/api/model/assignmentResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface AssignmentResponse {
  id?: string;
  courseId?: string;
  courseTitle?: string;
  title?: string;
  description?: string;
  submissionType?: string;
  isPublished?: boolean;
  publishAt?: string;
  deadlineAt?: string;
  maxScore?: number;
  languages?: string[];
  createdAt?: string;
  updatedAt?: string;
}